restyle step items & organize steplist comps
* move steplist comps into dir
* body1 font size for step items
* volume left-aligned & normal font color
