+++ protocol-designer/src/components/StepList/StepItem.js
@@ -3,10 +3,10 @@ import * as React from 'react'
 import cx from 'classnames'
 
 import {Icon, TitledList} from '@opentrons/components'
-import StepDescription from './StepDescription'
+import StepDescription from '../StepDescription'
 import styles from './StepItem.css'
 
-import {stepIconsByType, type StepType} from '../form-types'
+import {stepIconsByType, type StepType} from '../../form-types'
 
 type StepItemProps = {
   stepType: StepType,
